requirements.txt
# SecureAsk API Dependencies

# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0

# Graph Database
neo4j==5.18.0

# Cache
redis==5.0.1

# External APIs
aiohttp==3.9.3
beautifulsoup4==4.12.3

# Authentication
PyJWT==2.8.0

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1

# Monitoring
structlog==24.1.0